new non-cached pipeline
The pipeline for not-yet cached modules has been completely revamped.

Previously, when a cache miss happened for a package, qdd would download
the package and pipe the stream to a temporary file. This file would
then be untarred in a temporary directory. The contents would then be
copied to the cache, and then the contents of the cache would be copied
to the destination directory. That final step was the only one shared
with the cache-hit case.

With this change, qdd will pipe the http stream into `tar-stream`, which
calls a handler for each of the files. Inside the handler we have
readble streams containing the file data, so we just write those
directly to the cache directory and *simultaneously* to the destination
directory.

This new structure is greatly simplified, requires no temporary
files and has far fewer filesystem operations. As a result this is now
quite a bit faster. See the benchmark diff in README.md for exactly how
much faster.